+++ tests/Support/Constraint/CssConstraint.php
@@ -27,11 +27,15 @@ abstract class CssConstraint extends Constraint
      * @var string
      */
     private const CSS_REGULAR_EXPRESSIOM_PATTERN = '/
-        \\s*+([{},])\\s*+               # `{`, `}` or `,` captured in group 1, with possible whitespace either side
-        |(^\\s++)                       # whitespace at the very start, captured in group 2
-        |(>)\\s*+                       # `>` (e.g. closing a `<style>` element opening tag) with optional whitespace
-                                        # following, captured in group 3
-        |(?:(?!\\s*+[{},]|^\\s)[^>])++  # Anything else is matched, though not captured.  This is required so that any
+        (?<![\\s;}])\\s*+(\\})\\s*+     # `}`, captured in group 1, with possible whitespace either side, not preceded
+                                        # by `;` or `}`
+        |\\s*+([{};:,])\\s*+            # `{`, `}`, `;`, `:` or `,`, captured in group 2, with possible whitespace
+                                        # either side
+        |(^\\s++)                       # whitespace at the very start, captured in group 3
+        |(>)\\s*+                       # `>` (e.g. closing a `<style>` element opening tag), captured in group 4, with
+                                        # optional whitespace following
+        |(\\s++)                        # whitespace, captured in group 5
+        |(?:(?!\\s*+[{};:,]|\\s)[^>])++ # Anything else is matched, though not captured.  This is required so that any
                                         # characters in the input string that happen to have a special meaning in a
                                         # regular expression can be escaped.
     /x';
@@ -39,7 +43,7 @@ abstract class CssConstraint extends Constraint
     /**
      * Processing of @media rules may involve removal of some unnecessary whitespace from the CSS placed in the <style>
      * element added to the document, due to the way that certain parts are `trim`med.  Notably, whitespace either side
-     * of "{", "}" and "," or at the beginning of the CSS may be removed.
+     * of "{", "}", ";", ":", and "," or at the beginning of the CSS may be removed.
      *
      * This method helps takes care of that, by converting a search needle for an exact match into a regular expression
      * that allows for such whitespace removal, so that the tests themselves do not need to be written less humanly
@@ -56,7 +60,18 @@ abstract class CssConstraint extends Constraint
             [self::class, 'getCssNeedleRegularExpressionReplacement'],
             $needle
         );
-        return '/' . $needleMatcher . '/';
+        $needleMatcherAllowingQuotedUrl = \preg_replace(
+            '/url\\\\\\(((?:[^"\'\\s\\\\]|\\\\[^\\)])++)\\\\\\)/',
+            'url\\("?+$1"?+\\)',
+            $needleMatcher
+        );
+        $needleMatcherAllowingQuotedUrlOrMinifiedNumbers = \preg_replace(
+            ['/(\\.[1-9]++)0++(?!\\d)/', '/(\\d)\\\\\\.0++(?!\\d)/', '/(?<!\\d)0(\\\\\\.\\d)/'],
+            ['${1}0*+', '$1(?:\\\\.0++)?+', '0?+$1'],
+            $needleMatcherAllowingQuotedUrl
+        );
+
+        return '/' . $needleMatcherAllowingQuotedUrlOrMinifiedNumbers . '/';
     }
 
     /**
@@ -67,11 +82,15 @@ abstract class CssConstraint extends Constraint
     private static function getCssNeedleRegularExpressionReplacement(array $matches): string
     {
         if (($matches[1] ?? '') !== '') {
-            $regularExpressionEquivalent = '\\s*+' . \preg_quote($matches[1], '/') . '\\s*+';
+            $regularExpressionEquivalent = ';?+\\s*+' . \preg_quote($matches[1], '/') . '\\s*+';
         } elseif (($matches[2] ?? '') !== '') {
-            $regularExpressionEquivalent = '\\s*+';
+            $regularExpressionEquivalent = '\\s*+' . \preg_quote($matches[2], '/') . '\\s*+';
         } elseif (($matches[3] ?? '') !== '') {
-            $regularExpressionEquivalent = \preg_quote($matches[3], '/') . '\\s*+';
+            $regularExpressionEquivalent = '\\s*+';
+        } elseif (($matches[4] ?? '') !== '') {
+            $regularExpressionEquivalent = \preg_quote($matches[4], '/') . '\\s*+';
+        } elseif (($matches[5] ?? '') !== '') {
+            $regularExpressionEquivalent = '\\s++';
         } else {
             $regularExpressionEquivalent = \preg_quote($matches[0], '/');
         }