+++ tests/Unit/Support/Constraint/CssConstraintTest.php
@@ -25,11 +25,11 @@ final class CssConstraintTest extends TestCase
 
         $result = TestingCssConstraint::getCssNeedleRegularExpressionPatternForTesting($needle);
 
-        $resultWithWhitespaceMatchersRemoved = \str_replace('\\s*+', '', $result);
+        $resultWithOtherMatchersRemoved = \str_replace(['\\s*+', ';?+'], '', $result);
 
         self::assertSame(
             '/' . \preg_quote($needle, '/') . '/',
-            $resultWithWhitespaceMatchersRemoved
+            $resultWithOtherMatchersRemoved
         );
     }
 
@@ -39,7 +39,7 @@ final class CssConstraintTest extends TestCase
     public function getCssNeedleRegularExpressionPatternNotEscapesNonSpecialCharacters(): void
     {
         $needle = \implode('', \array_merge(\range('a', 'z'), \range('A', 'Z'), \range('0 ', '9 ')))
-            . "\r\n\t `¬\"£%&_;'@~,";
+            . '`¬"£%&_;\'@~,';
 
         $result = TestingCssConstraint::getCssNeedleRegularExpressionPatternForTesting($needle);
 
@@ -59,6 +59,8 @@ final class CssConstraintTest extends TestCase
         return [
             '"{" alone' => ['{', ''],
             '"}" alone' => ['}', ''],
+            '";" alone' => [';', ''],
+            '":" alone' => [':', ''],
             '"," alone' => [',', ''],
             '"{" with non-special character' => ['{', 'a'],
             '"{" with two non-special characters' => ['{', 'a0'],
@@ -85,10 +87,10 @@ final class CssConstraintTest extends TestCase
         );
 
         $quotedOtherContent = \preg_quote($otherContent, '/');
-        $expectedResult = '/' . $quotedOtherContent . '\\s*+' . \preg_quote($contentToInsertAround, '/') . '\\s*+'
-            . $quotedOtherContent . '/';
+        $expectedResult = $quotedOtherContent . '\\s*+' . \preg_quote($contentToInsertAround, '/') . '\\s*+'
+            . $quotedOtherContent;
 
-        self::assertSame($expectedResult, $result);
+        self::assertStringContainsString($expectedResult, $result);
     }
 
     /**